Made separate files for program.
